# source/rest-service/src/main/resources/static/api.yaml
openapi: 3.0.1
info:
  title: AERIUS emission service
  description: 'This is a service to retrieve AERIUS emission factors for different sectors'
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.nl.html
  version: 1.0.0
externalDocs:
  description: Find out more about AERIUS
  url: https://aerius.nl/
servers:
- url: http://localhost:8080/
tags:
- name: dataset
  description: Information about the different datasets
- name: sectors
  description: Sector specific information
- name: farming
  description: Farm specific information
paths:
  /datasets:
    get:
      summary: List all datasets
      operationId: listDatasets
      tags:
        - dataset
      parameters:
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: A list of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /datasets/{dataset}:
    get:
      summary: Retrieve information about a specific dataset
      operationId: getDataset
      tags:
        - dataset
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: The matching dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/sectors:
    get:
      summary: List all sectors
      operationId: listSectors
      tags:
        - sectors
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: A list of sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sector"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/farm/animals:
    get:
      summary: List all farm animal categories
      operationId: listFarmAnimals
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: A list of possible farm animal categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/farm/lodgings:
    get:
      summary: List all farm lodging categories
      operationId: listFarmLodgings
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
        - name: animalcode
          in: query
          description: filter for a specific animal category code.
          required: false
          schema:
            type: string
      responses:
        200:
          description: A list of possible farm lodging categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/{code}:
    get:
      summary: Get specific farm lodging information
      operationId: getFarmLodging
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
        - name: code
          in: path
          description: Which farm lodging category to retrieve.
          required: true
          schema:
            type: string
      responses:
        200:
          description: The matching farm lodging category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmLodgingCategory"
        404:
          description: Supplied dataset or category not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/additionalsystems:
    get:
      summary: List all farm additional lodging systems
      operationId: listFarmAdditionalLodgingSystems
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: A list of possible farm additional lodging systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/additionalsystems/{code}:
    get:
      summary: Get specific farm additional lodging system
      operationId: getFarmAdditionalLodgingSystem
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
        - name: code
          in: path
          description: Which farm additional lodging system category to retrieve.
          required: true
          schema:
            type: string
      responses:
        200:
          description: The matching farm additional lodging system category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmAdditionalLodgingSystemCategory"
        404:
          description: Supplied dataset or category not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/reductivesystems:
    get:
      summary: List all farm reductive lodging systems
      operationId: listFarmReductiveLodgingSystems
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: A list of possible farm reductive lodging systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/reductivesystems/{code}:
    get:
      summary: Get specific farm reductive lodging system
      operationId: getFarmReductiveLodgingSystem
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
        - name: code
          in: path
          description: Which farm reductive lodging system category to retrieve.
          required: true
          schema:
            type: string
      responses:
        200:
          description: The matching farm reductive lodging system category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmReductiveLodgingSystemCategory"
        404:
          description: Supplied dataset or category not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/foddermeasures:
    get:
      summary: List all farm lodging fodder measures
      operationId: listFarmLodgingFodderMeasures
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
      responses:
        200:
          description: A list of possible farm lodging fodder measures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        404:
          description: Supplied dataset not found.
          content: {}
  /datasets/{dataset}/farm/lodgings/foddermeasures/{code}:
    get:
      summary: Get specific farm lodging fodder measure
      operationId: getFarmLodgingFodderMeasure
      tags:
        - farming
      parameters:
        - $ref: '#/components/parameters/datasetParam'
        - $ref: '#/components/parameters/languageHeader'
        - name: code
          in: path
          description: Which farm lodging fodder measure category to retrieve.
          required: true
          schema:
            type: string
      responses:
        200:
          description: The matching farm lodging fodder measure category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmFodderMeasureCategory"
        404:
          description: Supplied dataset or category not found.
          content: {}
components:
  parameters:
    datasetParam:
      in: path
      name: dataset
      required: true
      schema:
        type: string
      description: Which dataset to retrieve data for.
    languageHeader:
      in: header
      name: Accept-Language
      required: false
      schema:
        type: string
      description:  Indication of which language to use. if not supplied or only unsupported languages are supplied, dutch will be used.
  schemas:
    Category:
      type: object
      description: Basic information for a category.
      required:
        - code
        - name
      properties:
        code:
          type: string
          description: Unique code for this category
        name:
          type: string
          description: The name for this category
        description:
          type: string
          description: The description for this category. Translated if possible.
    Dataset:
      allOf:
        - $ref: '#/components/schemas/Category'
        - description: A category for a sector.
        - type: object
          properties:
            metadata:
              type: object
              additionalProperties:
                type: string
    Sector:
      allOf:
        - $ref: '#/components/schemas/Category'
        - description: A category for a sector.
        - type: object
          properties:
            id:
              type: integer
              format: int64
              description: The ID of the sector. Deprecated in favor of code (which should be the string representation of this ID).
              deprecated: true
    FarmLodgingCategory:
      allOf:
        - $ref: '#/components/schemas/Category'
        - description: A category for a farm lodging type.
        - type: object
          required:
            - emissionFactors
            - systemDefinitions
            - additionalLodgingSystems
            - reductiveLodgingSystems
            - fodderMeasures
            - scrubber
          properties:
            emissionFactors:
              type: array
              description: Emission factors per animal.
              items:
                $ref: "#/components/schemas/EmissionFactor"
            systemDefinitions:
              type: array
              description: Possible farm lodging system definitions (BWL-number).
              items:
                $ref: "#/components/schemas/Category"
            additionalLodgingSystems:
              type: array
              description: Additional lodging systems that make sense to use for this type of farm lodging.
              items:
                $ref: "#/components/schemas/Category"
            reductiveLodgingSystems:
              type: array
              description: Reductive lodging systems that make sense to use for this type of farm lodging.
              items:
                $ref: "#/components/schemas/Category"
            fodderMeasures:
              type: array
              description: Fodder measures ('voermaatregelen') that make sense to use for this type of farm lodging.
              items:
                $ref: "#/components/schemas/Category"
            scrubber:
              type: boolean
              description: Indication wether or not this lodging type makes use of an air scrubber ('luchtwasser')
    FarmAdditionalLodgingSystemCategory:
      allOf:
        - $ref: '#/components/schemas/Category'
        - description: A category for a farm lodging additional system.
        - type: object
          required:
            - scrubber
            - emissionFactors
            - systemDefinitions
          properties:
            scrubber:
              type: boolean
              description: Indication wether or not this lodging type makes use of an air scrubber ('luchtwasser')
            emissionFactors:
              type: array
              description: The extra emission per animal that will be added to the lodging due to using this system.
              items:
                $ref: "#/components/schemas/EmissionFactor"
            systemDefinitions:
              type: array
              description: Possible farm lodging system definitions (BWL-number).
              items:
                $ref: "#/components/schemas/Category"
    FarmReductiveLodgingSystemCategory:
      allOf:
        - $ref: '#/components/schemas/Category'
        - description: A category for a farm lodging reductive system.
        - type: object
          required:
            - scrubber
            - reductionFractions
            - systemDefinitions
          properties:
            scrubber:
              type: boolean
              description: Indication wether or not this lodging type makes use of an air scrubber ('luchtwasser')
            reductionFractions:
              type: array
              description: The fraction that the total emission will be reduced to as a result of using this system.
              items:
                $ref: "#/components/schemas/EmissionReductionFraction"
            systemDefinitions:
              type: array
              description: Possible farm lodging system definitions (BWL-number).
              items:
                $ref: "#/components/schemas/Category"
    FarmFodderMeasureCategory:
      allOf:
        - $ref: '#/components/schemas/Category'
        - description: A category for a farm fodder measure.
        - type: object
          required:
            - reductionFractions
            - proportionsForAnimals
          properties:
            reductionFractions:
              type: array
              description: The different fractions per substance that are needed for emission calculations for this fodder measure.
              items:
                $ref: "#/components/schemas/FarmFodderMeasureEmissionFractions"
            proportionsForAnimals:
              type: array
              description: The proportions of emissions originating from the floor and the cellar per animal for this fodder measure type. 
              items:
                $ref: "#/components/schemas/FarmAnimalEmissionProportion"
    FarmFodderMeasureEmissionFractions:
      type: object
      required:
        - substance
        - totalReductionFraction
        - floorReductionFraction
        - cellarReductionFraction
      properties:
        substance:
          type: string
          description: The substance that the emission reduction fractions are for.
        totalReductionFraction:
          type: number
          description: The fraction that the emission will be reduced to. Should only be used if this is the only measure for the lodging. When there are multiple fodder measures the floor and cellar fractions in combination with the emission proportions for floor and cellar for the appropriate animal will have to be used to calculate the correct reductions of the combined fodder measures.
        floorReductionFraction:
          type: number
          description: The fraction that the emissions originating from the floor will be reduced to by using this system. 
        cellarReductionFraction:
          type: number
          description: The fraction that the emissions originating from the cellar (or manure pit) will be reduced to by using this system. 
    FarmAnimalEmissionProportion:
      type: object
      description: The proportion of emissions originating from the floor and the cellar per animal. The total of the fractions should sum up to be 1.
      required:
        - farmAnimalCode
        - proportionEmissionFromFloor
        - proportionEmissionFromCellar
      properties:
        farmAnimalCode:
          type: string
          description: The category code for the farm animal that these proportions are for.
        fractionFloor:
          type: number
          description: The fraction of the emission originating from the floor.
        fractionCellar:
          type: number
          description: The fraction of the emission originating from the cellar (or manure pit).
    EmissionFactor:
      type: object
      required:
        - substance
        - factor
      properties:
        substance:
          type: string
          description: The substance that the emission factor is for.
        factor:
          type: number
          description: The factor of emission per unit. Unit depends on the context where this object is used.
    EmissionReductionFraction:
      type: object
      required:
        - substance
        - fraction
      properties:
        substance:
          type: string
          description: The substance that the emission reduction fraction is for.
        fraction:
          type: number
          description: The fraction of emission.